chore: rename tree-view
